Colenso specimen and locality data
2:06 PM
This data was exported from EMu on 13/02/2015 and supplied to VUW
Wai-te-ata Press in response to a request for research data.

If you publish, distribute or otherwise disseminate this work to the
public without adapting it, the following attribution to Te Papa should
be used:

‘Source: Te Papa and licensed by Te Papa for re-use under the CC BY 3.0
NZ licence.

If you adapt this work in any way or include it in a collection, and
publish, distribute or otherwise disseminate that adaptation or
collection to the public, the following attribution to Te Papa should be
used:
‘This [work/product/application/etc] uses data sourced from Te Papa.’
For guidance on attribution please also see:
http://ict.govt.nz/guidance-and-resources/open-government/nzgoal/nzgoal-guidance-notes/nzgoal-guidance-note-4/data-specific-attribution-and-waiver-issues/
Please check the licence on the image at the page for the specimen. Most
images of Te Papa specimens are licenced CC BY-ND-NC
http://creativecommons.org/licenses/by-nc-nd/4.0/ but please confirm
before any reuse.
For further information please contact Adrian Kingston
adriank@tepapa.govt.nz or collectionsonline@tepapa.govt.nz